Extract globals and game session from omega.cpp
Move global variable definitions to globals.cpp and game initialization
plus the main loop to game_session.cpp, leaving omega.cpp as a thin
main() with signal handling. Split the game logic into an omega-core
static library that the omega executable links against.

init_game_session() now returns once the world is set up and drawn,
before the first turn runs; run_game_loop() takes a reset_clock flag
for new games. The call sequence is unchanged: time_clock(true) once
for a new game, then time_clock(false) repeatedly.

No functional change. This lets alternative front-ends (graphical
clients, tests) link the game core and drive the session themselves.
